requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.1
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
scikit-learn==1.2.2
python-dotenv==1.0.0
aiofiles==23.1.0
python-multipart==0.0.6